+++ src/Entities/GpApi/DTO/PaymentMethod.php
@@ -13,13 +13,12 @@ class PaymentMethod
     public $encryption;
     public $name;
     public $storage_mode;
-
     /** @var Card $card  */
     public $card;
-
     public $digital_wallet;
-
     public $bank_transfer;
+    /** @var array */
+    public $apm;
 
     const PAYMENT_METHOD_TOKEN_PREFIX = 'PMT_';
 }